pyproject.toml
[tool.poetry]
name = "sharded"
version = "0.0.1"
description = "Developer dependencies for sharded"
authors = ["Adam Garza <adriftible@gmail.com>"]
license = "MIT"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
rich = "^13.9.4"
appwrite = "^7.0.1"
discord-py = "^2.4.0"
python-dotenv = "^1.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
